+++ mailscanner/logout.php
@@ -4,7 +4,7 @@
  * MailWatch for MailScanner
  * Copyright (C) 2003-2011  Steve Freegard (steve@freegard.name)
  * Copyright (C) 2011  Garrod Alwood (garrod.alwood@lorodoes.com)
- * Copyright (C) 2014-2017  MailWatch Team (https://github.com/mailwatch/1.2.0/graphs/contributors)
+ * Copyright (C) 2014-2018  MailWatch Team (https://github.com/mailwatch/1.2.0/graphs/contributors)
  *
  * This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public
  * License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later
@@ -28,7 +28,7 @@
 require_once __DIR__ . '/functions.php';
 
 // Clear the login expiry timestamp
-$sql = "UPDATE users SET login_expiry='-1' WHERE username='" . safe_value($_SESSION['myusername']) . "'";
+$sql = "UPDATE users SET login_expiry='-1' WHERE username='" . safe_value(stripslashes($_SESSION['myusername'])) . "'";
 dbquery($sql);
 dbclose();