+++ mailscanner/mcp_rules_update.php
@@ -4,7 +4,7 @@
  * MailWatch for MailScanner
  * Copyright (C) 2003-2011  Steve Freegard (steve@freegard.name)
  * Copyright (C) 2011  Garrod Alwood (garrod.alwood@lorodoes.com)
- * Copyright (C) 2014-2017  MailWatch Team (https://github.com/mailwatch/1.2.0/graphs/contributors)
+ * Copyright (C) 2014-2018  MailWatch Team (https://github.com/mailwatch/1.2.0/graphs/contributors)
  *
  * This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public
  * License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later
@@ -71,7 +71,7 @@ if ($_SESSION['user_type'] !== 'A') {
                 $regs[2] = safe_value(ltrim(rtrim($regs[2])));
                 echo '<tr><td>' . htmlentities($regs[1]) . '</td><td>' . htmlentities($regs[2]) . '</td></tr>' . "\n";
                 dbquery("REPLACE INTO mcp_rules VALUES ('$regs[1]','$regs[2]')");
-                //debug("\t\tinsert: ".$regs[1].", ".$regs[2]);
+            //debug("\t\tinsert: ".$regs[1].", ".$regs[2]);
             } else {
                 debug("$line - did not match regexp, not inserting into database");
             }