+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+1.2.8
+-----------
+Improvements
+- Add ldap debug script
+- Update documentation
+- Improve sudo config to work with Postfix and Exim
+- Add mailwatch_update_sarules.php cronjob script
+- Use MAILWATCH_SMTP_HOSTNAME if defined on Release from quarantine (#1038)
+- Enable use of dn field as username field (#427, #1029)
+
+Fixes
+- Fix Postfix log processor (#1021)
+- Make msre reload script work on systemd
+- Fix error on undefined $_SERVER['HTTPS'] index
+- Fix logout process (#1044)
+- Fix email validation for username (#1042)
+
 1.2.7
 -----------
 Security